requirements.txt
ansible==6.0.0
jmespath==1.0.1
python-vagrant==1.0.0
pytest==7.1.2
pytest-testinfra==6.8.0
molecule==4.0.0
molecule-vagrant==1.0.0
